+++ packages/synthetix-main/test/integration/modules/core/CollateralModule/CollateralModuleDeposit.test.ts
@@ -2,7 +2,7 @@ import { ethers } from 'hardhat';
 import assert from 'assert/strict';
 import assertBn from '@synthetixio/core-utils/utils/assertions/assert-bignumber';
 import assertRevert from '@synthetixio/core-utils/utils/assertions/assert-revert';
-import { bootstrap } from '../bootstrap';
+import { bootstrap } from '../../../bootstrap';
 import { ethers as Ethers } from 'ethers';
 import assertEvent from '@synthetixio/core-utils/utils/assertions/assert-event';
